Add SLIM_SEO_DELETE_DATA constant to set whether the plugin deletes its data when uninstalling, default false.

+++ uninstall.php
@@ -1,10 +1,17 @@
 <?php
 defined( 'WP_UNINSTALL_PLUGIN' ) || die;
 
+// phpcs:ignore
+$delete_data = defined( 'SLIM_SEO_DELETE_DATA' ) ? SLIM_SEO_DELETE_DATA : false;
+if ( ! $delete_data ) {
+	return;
+}
+
 delete_option( 'slim_seo' );
 delete_option( 'slim_seo_db_version' );
 delete_option( 'ss_redirects' );
 delete_option( 'ss_redirection_db_version' );
 
 global $wpdb;
+// phpcs:ignore WordPress.DB.DirectDatabaseQuery
 $wpdb->query( "DROP TABLE IF EXISTS {$wpdb->prefix}slim_seo_404" );
